+++ inc/class-antivirus.php
@@ -722,13 +722,26 @@ class AntiVirus {
 
 								<p class="description">
 									<?php
-									/* translators: Link for transparency report */
-									$start_tag = sprintf( '<a href="%s">', esc_attr__( 'https://transparencyreport.google.com/safe-browsing/search?hl=en', 'antivirus' ) );
-									$end_tag = '</a>';
+									esc_html_e( 'Diagnosis and notification in suspicion case.', 'antivirus' );
+									echo ' ';
 									echo wp_kses(
-										/* translators: First placeholder (%1$s) starting link tag to transparency report, second placeholder (%2$s) closing link tag */
-										sprintf( __( 'Diagnosis and notification in suspicion case. For more details read %1$sthe transparency report%2$s.', 'antivirus' ), $start_tag, $end_tag ),
-										array( 'a' => array( 'href' => array() ) )
+										sprintf(
+											/* translators: First placeholder (%1$s) starting link tag to transparency report, second placeholder (%2$s) closing link tag */
+											__( 'For more details read %1$sthe transparency report%2$s.', 'antivirus' ),
+											sprintf(
+												'<a href="https://transparencyreport.google.com/safe-browsing/search?url=%s&hl=%s" target="_blank" rel="noopener noreferrer">',
+												urlencode( get_bloginfo( 'url' ) ),
+												substr( get_locale(), 0, 2 )
+											),
+											'</a>'
+										),
+										array(
+											'a' => array(
+												'href'   => array(),
+												'target' => array(),
+												'rel'    => array(),
+											),
+										)
 									);
 									?>
 								</p>