+++ antivirus.php
@@ -15,6 +15,7 @@
 
 /*
 Copyright (C)  2009-2015 Sergej Müller
+Copyright (C)  2016-2021 pluginkollektiv
 
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by