+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### unreleased ###
+* Point Safe Browsing link on settings page to site-specific URL (#106)
+
 ### 1.4.2 ###
 * Drop recursive check on option that failed in several scenarios (#96) (#97)
 * Drop check for base64 encoded strings which did not work properly in al cases (#100)