+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### unreleased ###
 * Point Safe Browsing link on settings page to site-specific URL (#106)
+* Increase the size of the Safe Browsing API input to show the entire key (#109)
 * Show warning if Safe Browsing check is enabled without custom API key (#105)
 
 ### 1.4.2 ###